% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer-gasmet.R
\name{analyzer_GASMET}
\alias{analyzer_GASMET}
\title{GASMET setup}
\usage{
analyzer_GASMET(manual_temperature = "temp", offset = "offset",
  duration_count = TRUE, plot = "plot", date = "date",
  start = "start", end = 10, trimmer = NULL, V = 0.01461,
  A = 0.098)
}
\arguments{
\item{manual_temperature}{Column conatining an additional measured temperature.
If specified, the `temperature` value is ommited.}

\item{offset}{(Default = `0:0`). Colon separated value for offsetting the
datasets. First value offsets the first, second value offsets the last
datapoint. Positive values shrinken the dataset i.e `2:3` skips two
datapoints at the beginning and truncates three datapoints at the end of
the dataset. Negative values expand the dataset i.e. `-2:-3` add two
datapoints at the beginning and three datapoints at the end of the dataset.}

\item{duration_count}{If TRUE, `end` is determined by the number of datapoints
within each measurement. If FALSE, `end` is treated as duration or point in time.}

\item{plot}{(Default = "plot"). Column containing the ID of the measurement location.}

\item{date}{(Default = "date"). Column containing the day of measurement. Format: "YYYY-MM-DD".}

\item{start}{(Default = "start"). Column containing the start of measurement. Format: "HH:MM" or "HH:MM:SS".}

\item{end}{(Default = "end"). Column containing the end of measurement. There are several options:

  Point in Time. Format: "HH:MM" or "HH:MM:SS" (e.g. `"13:43"`).

  Duration in minutes (e.g. `5`).

  Number of Datapoints regardless of the effectively elapsed time (e.g. `10`).
  This behaviour is controlled by `duration_count`.

  Alternatively a single number treated as Duration or Number of Datapoints for all measurements (e.g. `5`, behaviour is again controlled by `duration_count`).}

\item{trimmer}{A function to trim the intervall between `start` and `end`. See `trim_time()` for details.}

\item{V}{Volume of the chamber.}

\item{A}{Area of the chamber.}
}
\description{
Explicitly create GASMET setup. Usually you will not call this
  function, but will instead use the wrapper \code{\link{process_gasmet}}
}
