% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_shortcuts.R
\name{axis.flux.ch4}
\alias{axis.flux.ch4}
\title{Label axis for CH4 according to flux scale.}
\usage{
axis.flux.ch4(time)
}
\arguments{
\item{time}{character; Timescale. See 'Details'.}
}
\description{
Label axis for CH4 according to flux scale.
}
