% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_shortcuts.R
\name{axis.flux.co2}
\alias{axis.flux.co2}
\title{Label axis for CO2 according to flux scale.}
\usage{
axis.flux.co2(time)
}
\arguments{
\item{time}{character; Timescale. See 'Details'.}
}
\value{
ggplot ylab object.
}
\description{
Label axis for CO2 according to flux scale.
}
