% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_shortcuts.R
\name{axis.flux.gas}
\alias{axis.flux.gas}
\title{Label axis for arbitrary gases according to flux scale.}
\usage{
axis.flux.gas(time)
}
\arguments{
\item{time}{character; Timescale. See 'Details'.}
}
\description{
Label axis for arbitrary gases according to flux scale.
}
