% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc.kW}
\alias{calc.kW}
\title{Calculate Gas transfer coefficient}
\usage{
calc.kW(ws, wt, gas = "co2", model = "wann14")
}
\arguments{
\item{ws}{numeric; windspeed m/s}

\item{wt}{numeric; water temperature °C}

\item{gas}{character; one of 'CO2', 'CH4'}

\item{model}{character; model to calculate kW. See 'Details'}
}
\value{
kW for the given parameters
}
\description{
Calculate Gas transfer coefficient
}
\details{
Currently these models are supported:

  WANN14: Wanninkhof (2014)
}
\references{
Wanninkhof, R., 2014. Relationship between wind speed and gas exchange over the ocean revisited. Limnology and Oceanography: Methods 12, 351–362. https://doi.org/10.4319/lom.2014.12.351
}
