% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_kW}
\alias{calc_kW}
\title{Returns the gas transfer coefficient for gas of interest based on different
models.}
\usage{
calc_kW(ws, wt, gas = "CO2", model = "wann14")
}
\arguments{
\item{ws}{numeric; windspeed m/s}

\item{wt}{numeric; water temperature °C}

\item{gas}{character; one of 'CO2', 'CH4'}

\item{model}{character; model to calculate kW. See 'Details'}
}
\value{
Numeric value of gas transfer coefficient for the given parameters
}
\description{
Returns the gas transfer coefficient for gas of interest based on different
models.
}
\details{
Currently these models are supported:

 \itemize{
   \item WANN14: Wanninkhof (2014)
   \item COLE98: Cole and Caraco (1998)
 }
}
\references{
Wanninkhof, R., 2014. Relationship between wind speed and gas
exchange over the ocean revisited. Limnology and Oceanography: Methods 12,
351–362. https://doi.org/10.4319/lom.2014.12.351

  Cole, J.J., Caraco, N.F., 1998. Atmospheric exchange of carbon dioxide in
  a low-wind oligotrophic lake measured by the addition of SF6. Limnology
  and Oceanography 43, 647–656. https://doi.org/10.4319/lo.1998.43.4.0647
}
