% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_SN}
\alias{calc_SN}
\title{Calculate Schmidt Number}
\usage{
calc_SN(wt, gas = "co2")
}
\arguments{
\item{wt}{numeric; Water temperature.}

\item{gas}{character; Gas to calculate SN for. See 'Details'.}
}
\value{
test
}
\description{
Calculate Schmidt Number
}
