% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals}
\alias{gals}
\title{Gas analyzer setup}
\usage{
gals(..., .default = FALSE)
}
\arguments{
\item{...}{List of name value pairs to setup the gasanalyzer.}

\item{.default}{If TRUE, default settings are used in addition to the passed
ones. Useful to abbreviate the function call.}
}
\value{
A gals object.
}
\description{
Gas analyzer setup
}
\section{Definition}{

custom uses the following parameters (fixed parameters are in bold):
\itemize{
  \item name = custom
  \item time_stamp = 
  \item conc_columns = 
  \item preassure = 
  \item preassure_factor = 1
  \item temperature = 
  \item manual_temperature = 
  \item offset = 0
  \item duration_count = FALSE
  \item spot = spot
  \item day = day
  \item start = start
  \item end = end
  \item trimmer = 
  \item V = 0.01461
  \item A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

