% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals_gasmet}
\alias{gals_gasmet}
\title{GASMET setup}
\usage{
gals_gasmet(.default = TRUE)
}
\description{
Explicitly create LosGatos setup. Usually you will not call this
  function, but will instead use the wrapper [process_gasmet()]
}
\section{Definition}{

gasmet uses the following parameters (fixed parameters are in bold):
\itemize{
  \item \strong{name} = gasmet
  \item \strong{time_stamp} = datetime
  \item \strong{conc_columns} = CH4, CO2, N2O
  \item \strong{preassure} = Luftdruck
  \item \strong{preassure_factor} = 1
  \item \strong{temperature} = NA
  \item manual_temperature = temp
  \item offset = 0
  \item \strong{duration_count} = TRUE
  \item spot = spot
  \item day = day
  \item start = start
  \item end = wndw
  \item trimmer = 
  \item V = 0.01461
  \item A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

