% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals_losgatos}
\alias{gals_losgatos}
\title{LosGatos setup}
\usage{
gals_losgatos(.default = TRUE)
}
\description{
Explicitly create LosGatos setup. Usually you will not call this
function, but will instead use the wrapper \link[=process_chamber]{process_losgatos()}.
}
\section{Definition}{

losgatos uses the following parameters (fixed parameters are in bold):
\preformatted{
name = losgatos
time_stamp = Time
conc_columns = [CH4]_ppm, [CO2]_ppm
preassure = GasP_torr
preassure_factor = 1.33322
temperature = AmbT_C
manual_temperature = 
offset = 0
duration_count = FALSE
spot = spot
day = day
start = start
end = end
trimmer = trim_time
V = 0.01461
A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

