% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasmet}
\alias{gasmet}
\title{GASMET Flux measurement with three entrys.}
\format{A data frame with 25 rows and 7 variables:
\describe{
  \item{Datum}{Date of measurement}
  \item{Zeit}{Time of the datapoint}
  \item{CO2}{}
  \item{CH4}{}
  \item{N2O}{}
  \item{Luftdruck}{}
  \item{datetime}{}
}}
\usage{
gasmet
}
\description{
A dataset containing three flux measurements.
}
\keyword{datasets}
