% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmr.R
\docType{package}
\name{glimmr-package}
\alias{glimmr}
\alias{glimmr-package}
\title{glimmr: Gas fLuxes and dynamIc chaMber MeasuRements}
\description{
The glimmr package provides functions to convert high frequency concentration data
obtained by chamber measurements into gasfluxes. Data recorded with GASMET and LosGatos devices can be
processed automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tekknosol/glimmr}
  \item Report bugs at \url{https://github.com/tekknosol/glimmr/issues}
}

}
\author{
\strong{Maintainer}: Philipp S. Keller \email{phlipp.keller@ufz.de}

}
\keyword{internal}
