% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{inspect_gasmet}
\alias{inspect_gasmet}
\alias{inspect_losgatos}
\title{Show plots for every measurement}
\usage{
inspect_gasmet(fluxdata, meta, manual_temperature = "temp", spot = "spot",
  day = "day", start = "start", end = "wndw")

inspect_losgatos(fluxdata, meta, manual_temperature = NA, spot = "spot",
  day = "day", start = "start", end = "end")
}
\arguments{
\item{fluxdata}{A data frame with recorded data. Either loaded with
\code{\link{read_gasmet}} or \code{\link{read_losgatos}}.}

\item{meta}{Data frame. Metadata containing required informations.}

\item{manual_temperature}{Character.}

\item{spot}{Character.}

\item{day}{Character.}

\item{start}{Character.}

\item{end}{Character.}
}
\value{
A series of plots
}
\description{
Show plots for every measurement
}
