% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{kH}
\alias{kH}
\title{Calculate Henry's constant}
\usage{
kH(t, gas = "co2")
}
\arguments{
\item{t}{numeric; temperature °C}

\item{gas}{character; Gas to calculate kH for. See 'Details'.}
}
\value{
Henry's constant
}
\description{
Calculate Henry's constant
}
\details{
Currently these gases are supported:

 \itemize{
   \item CO_2
   \item CH_4
 }
}
