% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{losgatos}
\alias{losgatos}
\title{LosGatos Flux measurement with three entries.}
\format{A data frame with 1684 rows and 5 variables:
\describe{
  \item{Time}{Datetime of measurement}
  \item{[CH4]_ppm}{}
  \item{[CO2]_ppm}{}
  \item{GasP_torr}{}
  \item{AmbT_C}{}
}}
\usage{
losgatos
}
\description{
A dataset containing one flux measurements.
}
\keyword{datasets}
