% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_gasmet}
\alias{meta_gasmet}
\title{Metafile for GASMET measurement.}
\format{A data frame with 3 rows and 19 variables:
\describe{
  \item{spot}{ID of the measurement spot}
  \item{day}{Day of the measurement}
  \item{begin}{}
  \item{temp}{}
  \item{wndw}{}
  \item{offset}{}
}}
\usage{
meta_gasmet
}
\description{
A dataset containing the meta information for the three flux measurements.
}
\keyword{datasets}
