% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_losgatos}
\alias{meta_losgatos}
\title{Metafile for LosGatos measurement.}
\format{A data frame with 1 rows and 19 variables:
\describe{
  \item{plot}{ID of the measurement spot}
  \item{date}{Day of the measurement}
  \item{start}{}
  \item{end}{}
  \item{t_start}{}
  \item{t_end}{}
  \item{offset}{}
}}
\usage{
meta_losgatos
}
\description{
A dataset containing the meta information for one flux measurements.
}
\keyword{datasets}
