% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{preprocess_gasmet}
\alias{preprocess_gasmet}
\title{Preprocess GASMET data for flux calculation}
\usage{
preprocess_gasmet(gasmet, meta, V = 0.01461, A = 0.098)
}
\arguments{
\item{gasmet}{data frame; Recordings from GASMET.}

\item{meta}{data frame; Metainformations, see 'Details'.}

\item{V}{numeric; Chamber Volume.}

\item{A}{numeric; Chamber Area.}
}
\description{
Preprocess GASMET data for flux calculation
}
