% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{process_losgatos}
\alias{process_losgatos}
\title{Process flux calculation for LosGatos data.}
\usage{
process_losgatos(losgatos, meta, V = 0.01461, A = 0.098, pre = F)
}
\arguments{
\item{losgatos}{a data frame containing LosGatos data.}

\item{meta}{a data frame of meta information. See 'Details'.}

\item{V}{numeric; Chamber volume.}

\item{A}{numeric; Chamber area.}

\item{pre}{if TRUE don't process the flux calculation and return the
preprocessed dataframe. When FALSE (the default) data is processed with
routines from 'gasfluxes' package and the fluxes are returned.}
}
\value{
Fluxes
}
\description{
Process flux calculation for LosGatos data.
}
