% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{read_gasmet}
\alias{read_gasmet}
\title{Read a GASMET file.}
\usage{
read_gasmet(file)
}
\arguments{
\item{file}{Path to a file}
}
\value{
A data frame.
}
\description{
Read a GASMET file.
}
\examples{
\dontrun{
 gasmet <- read_gasmet("path/to/gasmet.txt")
}

}
