% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{read_losgatos}
\alias{read_losgatos}
\title{Read a LosGatos file.}
\usage{
read_losgatos(file, ...)
}
\arguments{
\item{file}{Path to a file.}

\item{...}{additional parameters passed to read_csv()}
}
\value{
A data frame.
}
\description{
Read a LosGatos file.
}
