% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sec2day}
\alias{sec2day}
\title{Converts µMol m⁻² d⁻¹ to mmol m⁻² d⁻¹}
\usage{
sec2day(flux)
}
\arguments{
\item{flux}{numeric; Flux value in µMol m⁻² d⁻¹}
}
\value{
Flux in mmol m⁻² d⁻¹
}
\description{
Converts µMol m⁻² d⁻¹ to mmol m⁻² d⁻¹
}
