% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tregata.R
\docType{package}
\name{tregata-package}
\alias{tregata}
\alias{tregata-package}
\title{tregata: (Meta-)Package for the TregaTa project}
\description{
The tregata package provides functions to convert high frequency concentration data
obtained by chamber measurements into gasfluxes. Data recorded with GASMET and LosGatos devices can be
processed automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tekknosol/tregata}
  \item Report bugs at \url{https://github.com/tekknosol/tregata/issues}
}

}
\author{
\strong{Maintainer}: Philipp Keller \email{phlipp.keller@ufz.de}

}
\keyword{internal}
