% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tregata.R
\docType{package}
\name{tregata}
\alias{tregata}
\alias{tregata-package}
\title{tregata: A package for waterchemestry, limnophysics and gasflux calculations.}
\description{
The tregata package provides functions to convert high frequency concentration data
obtained by chamber measurements into gasfluxes. Data recorded with GASMET and LosGatos devices can be
processed automatically.
}
