% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{trimmers}
\alias{trimmers}
\alias{trim_time}
\title{Functions to modify interval between start and end of chamber deployment}
\usage{
trim_time(interval)
}
\arguments{
\item{interval}{Intervall of chamber deployment. Created with
lubridate::intervall()}
}
\value{
Modified interval
}
\description{
These functions are used to modify the measurement interval.
}
\details{
`trim_time()` increases start of interval to the nearest minute and
 decreases end of interval to the nearest minute.
}
\examples{
library(lubridate)
start <- ymd_hms("2018-06-25 12:13")
end <- ymd_hms("2018-06-25 12:18")
int <- interval(start, end)

trim_time(int)
}
