% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_shortcuts.R
\name{axis.flux.co2}
\alias{axis.flux.co2}
\alias{axis.flux.gas}
\alias{axis.flux.ch4}
\title{Label ggplot2 axis according to temporal scale of gasflux}
\usage{
axis.flux.co2(time)

axis.flux.gas(time)

axis.flux.ch4(time)
}
\arguments{
\item{time}{character; Either "sec" oder "day". See 'Details'.}
}
\value{
Returns \code{\link[ggplot2]{ylab}} object
}
\description{
Label ggplot2 axis according to temporal scale of gasflux
}
\details{
"sec" returns expression(## ~ flux ~ ("µ" ~ mol ~ m^-2 ~ s^-1))),

    "day" returns expression(## ~ flux ~ (mmol ~ m^-2 ~ d^-1))),

    where ## is either "Gas", "\out{CO<sub>2</sub>}", or "\out{CH<sub>4</sub>}"
}
\examples{
library(ggplot2)

df <- data.frame(
    plot = rep(c("A", "B", "C"), each = 20),
    flux = c(rnorm(20), rnorm(20), rnorm(20))
)

ggplot(df, aes(plot, flux))+
geom_boxplot()+
axis.flux.gas("day")

}
