% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc.kW}
\alias{calc.kW}
\title{calc.kW}
\usage{
calc.kW(ws, wt, gas = "co2", model = "wann14")
}
\arguments{
\item{ws}{numeric; windspeed m/s}

\item{wt}{numeric; water temperature °C}

\item{gas}{character; one of 'CO2', 'CH4'}

\item{model}{character; model to calculate}
}
\value{
kW for the given parameters
}
\description{
calc.kW
}
