% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_kH}
\alias{calc_kH}
\title{Returns Henry's constant for gas of interest}
\usage{
calc_kH(t, gas = "CO2")
}
\arguments{
\item{t}{numeric; temperature in °C}

\item{gas}{character; Gas to calculate kH for. See 'Details'.}
}
\value{
Numeric value of Henry's constant \out{H<sup>cp</sup>} (mol/(L*atm))
}
\description{
Returns Henry's constant for gas of interest
}
\details{
Currently supported gases:

 \itemize{
   \item CO2
   \item CH4
 }
}
\examples{
watertemperature <- rnorm(5, mean = 20)

calc_kH(watertemperature)

calc_kH(watertemperature, "ch4")

}
