% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_SN}
\alias{calc_SN}
\title{Calculate Schmidt Number}
\usage{
calc_SN(wt, gas = "CO2")
}
\arguments{
\item{wt}{numeric; Water temperature.}

\item{gas}{character; Gas to calculate SN for. See 'Details'.}
}
\value{
test
}
\description{
Calculate Schmidt Number
}
\examples{
watertemperature <- rnorm(5, mean = 20)

calc_SN(watertemperature)

calc_SN(watertemperature, "ch4")
}
