% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{day2sec}
\alias{day2sec}
\title{Convert mmol m⁻² d⁻¹ to µMol m⁻² d⁻¹}
\usage{
day2sec(flux)
}
\arguments{
\item{flux}{numeric; Flux value in mmol m⁻² d⁻¹}
}
\value{
Flux in µMol m⁻² d⁻¹
}
\description{
Convert mmol m⁻² d⁻¹ to µMol m⁻² d⁻¹
}
