% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{day2sec}
\alias{day2sec}
\alias{sec2day}
\title{Performs unit conversion for gas fluxes}
\usage{
day2sec(flux)

sec2day(flux)
}
\arguments{
\item{flux}{numeric; Flux value in mmol/sqm/d or µMol/sqm/d}
}
\value{
Numeric value of converted flux.
}
\description{
Performs unit conversion for gas fluxes
}
