% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals}
\alias{gals}
\title{Gas analyzer setup}
\usage{
gals(name = "custom", fixed_params = c(), time_stamp = NULL,
  conc_columns = NULL, preassure = NULL, preassure_factor = 1,
  temperature = NULL, manual_temperature = NULL, duration_count = FALSE,
  spot = "spot", day = "day", start = "start", end = "end",
  trimmer = NULL, V = 0.01461, A = 0.098)
}
\arguments{
\item{name}{Name of device. Defaults to "custom"}

\item{fixed_params}{Named vector.}

\item{time_stamp}{Character vector. Value will be used as the column in conc
containing the timestamp.}

\item{conc_columns}{A named character vector. Values will be used to extract
ppm data from conc. E.g. c(CO2 = "CO2").}

\item{preassure}{Character}

\item{preassure_factor}{Numeric value to convert preassure to mbar.}

\item{temperature}{Character}

\item{manual_temperature}{Character}

\item{duration_count}{Is `end` given as time information or number of
datapoints? If FALSE (the default), `end` will be interpreted as Time or
duration in minutes. If TRUE, `end` will be interpreted as number of
datapoints per chamber application.}

\item{spot}{Character.}

\item{day}{Character.}

\item{start}{Character.}

\item{end}{Character.}

\item{trimmer}{NULL or function. Function is used to modify interval between
`start` and `end`.}

\item{V}{Numeric. Volume of used chamber.}

\item{A}{Numeric. Area of used chamber.}
}
\value{
A gals object.
}
\description{
Gas analyzer setup
}
