% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals}
\alias{gals}
\title{Gas analyzer setup}
\usage{
gals(..., .default = FALSE)
}
\arguments{
\item{...}{List of name value pairs to setup the gasanalyzer.}

\item{.default}{If TRUE, default settings are used in addition to the passed
ones. Useful to abbreviate the function call.}
}
\value{
A gals object.
}
\description{
Gas analyzer setup
}
\section{Definition}{

custom uses the following parameters (fixed parameters are in bold):
\preformatted{
name = custom
time_stamp = 
conc_columns = 
preassure = 
preassure_factor = 1
temperature = 
manual_temperature = 
offset = 0:0
duration_count = FALSE
spot = spot
day = day
start = start
end = end
trimmer = 
V = 0.01461
A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

