% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals_gasmet}
\alias{gals_gasmet}
\title{GASMET setup}
\usage{
gals_gasmet(.default = TRUE)
}
\description{
Explicitly create LosGatos setup. Usually you will not call this
  function, but will instead use the wrapper [process_gasmet()]
}
\section{Definition}{

gasmet uses the following parameters (fixed parameters are in bold):
\preformatted{
name = gasmet
time_stamp = datetime
conc_columns = CH4, CO2, N2O
preassure = Luftdruck
preassure_factor = 1
temperature = NA
manual_temperature = temp
offset = 0
duration_count = TRUE
spot = spot
day = day
start = start
end = wndw
trimmer = 
V = 0.01461
A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

