% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals.R
\name{gals_losgatos}
\alias{gals_losgatos}
\title{LosGatos setup}
\usage{
gals_losgatos(.default = TRUE)
}
\description{
Explicitly create LosGatos setup. Usually you will not call this
function, but will instead use the wrapper [process_losgatos()]
}
\section{Definition}{

losgatos uses the following parameters (fixed parameters are in bold):
\itemize{
  \item \strong{name} = losgatos
  \item \strong{time_stamp} = Time
  \item \strong{conc_columns} = [CH4]_ppm, [CO2]_ppm
  \item \strong{preassure} = GasP_torr
  \item \strong{preassure_factor} = 1.33322
  \item \strong{temperature} = AmbT_C
  \item manual_temperature = 
  \item offset = 0
  \item duration_count = FALSE
  \item spot = spot
  \item day = day
  \item start = start
  \item end = end
  \item trimmer = trim_time
  \item V = 0.01461
  \item A = 0.098
}
Learn more about these definitions in \code{vignette("gals")}
}

