% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmr.R
\docType{package}
\name{glimmr-package}
\alias{glimmr}
\alias{glimmr-package}
\title{glimmr: Gas Fluxes and Dynamic Chamber Measurements}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

provides functions to convert high frequency concentration data
    obtained by chamber measurements into gasfluxes. Data recorded with GASMET
    and LosGatos devices can be processed automatically. Furthermore, functions
    for flux calculation using BLE are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tekknosol.github.io/glimmr/}
  \item \url{https://github.com/tekknosol/glimmr}
  \item Report bugs at \url{https://github.com/tekknosol/glimmr/issues}
}

}
\author{
\strong{Maintainer}: Philipp S. Keller \email{phlipp.keller@ufz.de} (0000-0002-3565-4413)

}
\keyword{internal}
