% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\docType{data}
\name{inspect_gasmet}
\alias{inspect_gasmet}
\alias{inspect_losgatos}
\alias{CO2mmol}
\title{Show plots for every measurement}
\format{An object of class \code{NULL} of length 0.}
\usage{
inspect_gasmet(fluxdata, meta)

CO2mmol
}
\arguments{
\item{fluxdata}{A data frame with recorded data. Either loaded with
\code{\link{read_gasmet}} or \code{\link{read_losgatos}}.}

\item{meta}{Metadata}
}
\value{
A series of plots
}
\description{
Show plots for every measurement
}
\keyword{datasets}
