% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{inspect_gasmet}
\alias{inspect_gasmet}
\alias{inspect_losgatos}
\alias{inspect_losgatos}
\title{Show plots for every measurement}
\usage{
inspect_gasmet(gasmet, meta)

inspect_losgatos(losgatos, meta)
}
\arguments{
\item{gasmet}{A GASMET data frame}

\item{meta}{Metadata}
}
\value{
A series of plots
}
\description{
Show plots for every measurement
}
