% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{inspect_gasmet}
\alias{inspect_gasmet}
\alias{inspect_losgatos}
\alias{inspect_losgatos}
\title{Show plots for every measurement}
\usage{
inspect_gasmet(fluxdata, meta)

inspect_losgatos(fluxdata, meta)
}
\arguments{
\item{fluxdata}{A data frame with recorded data. Either loaded with
\code{\link{read_gasmet}} or \code{\link{read_losgatos}}.}

\item{meta}{Metadata}
}
\value{
A series of plots
}
\description{
Show plots for every measurement
}
