% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{preprocess_gasmet}
\alias{preprocess_gasmet}
\title{Preprocess GASMET data for flux calculations}
\usage{
preprocess_gasmet(gasmet, meta, V = 0.01461, A = 0.098)
}
\arguments{
\item{gasmet}{data frame; Recordings from GASMET.}

\item{meta}{data frame; Metainformations, see 'Details'.}

\item{V}{numeric; Chamber Volume. Default to 0.01461.}

\item{A}{numeric; Chamber Area. Default to 0.098.}
}
\value{
dataframe which can be used as input of the gasfluxes function
}
\description{
Preprocess GASMET data for flux calculations
}
