% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{process_chamber}
\alias{process_chamber}
\alias{process_losgatos}
\alias{process_gasmet}
\title{Calculate gasfluxes from dynamic chamber measurement}
\usage{
process_chamber(conc, meta, time_stamp, conc_columns, preassure,
  preassure_factor = 1, temperature = "temp", manual_temperature = NA,
  duration_count = FALSE, spot = "spot", day = "day", start = "start",
  end = "end", V, A, time_proc = NA, pre = FALSE)

process_losgatos(data, meta, manual_temperature = NA, spot = "spot",
  day = "day", start = "start", end = "end", V = 0.01461, A = 0.098,
  pre = FALSE)

process_gasmet(data, meta, manual_temperature = "temp", spot = "spot",
  day = "day", start = "start", end = "wndw", V, A, pre = FALSE)
}
\arguments{
\item{conc}{Data frame. Recorded data from device.}

\item{meta}{Data frame. Metadata containing required informations.}

\item{time_stamp}{Character vector. Value will be used as the column in conc
containing the timestamp.}

\item{conc_columns}{A named character vector. Values will be used to extract
ppm data from conc. E.g. c(CO2 = "CO2").}

\item{preassure}{Character.}

\item{preassure_factor}{Numeric value to convert preassure to mbar.}

\item{temperature}{Character.}

\item{duration_count}{Is `end` given as time information or number of
datapoints? If FALSE (the default), `end` will be interpreted as Time or duration in
minutes. If TRUE, `end` will be interpreted as number of datapoints per
chamber application.}

\item{spot}{Character.}

\item{day}{Character.}

\item{start}{Character.}

\item{end}{Character.}

\item{V}{Numeric. Volume of used chamber.}

\item{A}{Numeric. Area of used chamber.}

\item{time_proc}{NA or name of function. Function is used to modify interval
between `start` and `end`.}

\item{pre}{If FALSE, flux processing will be executed. If FALSE flux
processing will be skipped and preprocessed data frame will be returned.}

\item{manuel_temperature}{Character.}
}
\value{
A data frame.
}
\description{
`process_losgatos()`and `process_gasmet()` are special cases of the general
`process_chamber()`.
}
