% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{process_chamber}
\alias{process_chamber}
\alias{process_losgatos}
\alias{process_gasmet}
\title{Calculate gasfluxes from dynamic chamber measurement}
\usage{
process_chamber(data = NULL, meta = NULL, analyzer = gals(), pre = T)

process_losgatos(data, meta, analyzer = NULL, pre = FALSE)

process_gasmet(data, meta, analyzer = NULL, pre = FALSE)
}
\arguments{
\item{data}{Data frame. Recorded data from device.}

\item{meta}{Data frame. Metadata containing required informations.}

\item{analyzer}{Set of setup mappings created by [gals()]. The specified name
value pairs will override default settings.}

\item{pre}{If FALSE (the default), flux processing will be executed. If TRUE
flux processing will be skipped and preprocessed data frame will be
returned.}
}
\value{
A data frame.
}
\description{
`process_losgatos()`and `process_gasmet()` are special cases of the general
`process_chamber()` with preconfigured settings. See [gals_losgatos()] and
[gals_gasmet()] for details.
}
