% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{read_losgatos}
\alias{read_losgatos}
\title{Read a LosGatos file.}
\usage{
read_losgatos(path, clean_dir = FALSE, ...)
}
\arguments{
\item{clean_dir}{If TRUE directory is cleaned}

\item{...}{additional parameters passed to read_csv()}

\item{file}{Path to a file or directory.}
}
\value{
A data frame.
}
\description{
Read a LosGatos file.
}
