% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_utility.R
\name{read_losgatos}
\alias{read_losgatos}
\title{Read a LosGatos file.}
\usage{
read_losgatos(path, format = "dmy", clean_dir = FALSE, ...)
}
\arguments{
\item{path}{Path to a file or directory.}

\item{format}{either "dmy" or "mdy".}

\item{clean_dir}{If TRUE directory is cleaned}

\item{...}{additional parameters passed to read_csv()}
}
\value{
A data frame.
}
\description{
Read a LosGatos file.
}
